Convert ratings page to composable
